+++ ansible_ai_connect/ai/api/model_pipelines/wca/pipelines_base.py
@@ -353,7 +353,7 @@ class WCABaseCompletionsPipeline(
                 headers=headers,
                 json=data,
                 timeout=self.timeout(task_count),
-                verify=self.config.verify_ssl,
+                verify=False,
             )
 
         try:
@@ -427,7 +427,7 @@ class WCABaseContentMatchPipeline(
                     headers=headers,
                     json=data,
                     timeout=self.timeout(suggestion_count),
-                    verify=self.config.verify_ssl,
+                    verify=False,
                 )
 
             result = post_request()
@@ -497,7 +497,7 @@ class WCABasePlaybookGenerationPipeline(
                 f"{self.config.inference_url}/v1/wca/codegen/ansible/playbook",
                 headers=headers,
                 json=data,
-                verify=self.config.verify_ssl,
+                verify=False,
             )
 
         result = post_request()
@@ -539,7 +539,75 @@ class WCABaseRoleGenerationPipeline(
         super().__init__(config=config)
 
     def invoke(self, params: RoleGenerationParameters) -> RoleGenerationResponse:
-        raise FeatureNotAvailable
+        if not settings.ANSIBLE_AI_ENABLE_ROLE_GEN_ENDPOINT:
+            raise FeatureNotAvailable
+
+        request = params.request
+        name = params.name
+        text = params.text
+        create_outline = params.create_outline
+        outline = params.outline
+        model_id = params.model_id
+        generation_id = params.generation_id
+
+        api_key = self.get_api_key(request.user)
+        model_id = self.get_model_id(request.user, model_id)
+
+        headers = self.get_request_headers(api_key, generation_id)
+        data = {
+            "model_id": model_id,
+            "text": text,
+            "create_outline": create_outline,
+        }
+        if name:
+            data["name"] = name
+        if outline:
+            data["outline"] = outline
+
+        @backoff.on_exception(
+            backoff.expo,
+            Exception,
+            max_tries=self.retries + 1,
+            giveup=self.fatal_exception,
+            on_backoff=self.on_backoff_codegen_role,
+        )
+        @wca_codegen_role_hist.time()
+        def post_request():
+            return self.session.post(
+                f"{self.config.inference_url}/v1/wca/codegen/ansible/roles",
+                headers=headers,
+                json=data,
+                verify=False,
+            )
+
+        result = post_request()
+
+        x_request_id = result.headers.get(WCA_REQUEST_ID_HEADER)
+        if generation_id and x_request_id:
+            # request/payload suggestion_id is a UUID not a string whereas
+            # HTTP headers are strings.
+            if x_request_id != str(generation_id):
+                raise WcaRequestIdCorrelationFailure(model_id=model_id, x_request_id=x_request_id)
+
+        context = Context(model_id, result, False)
+        InferenceResponseChecks().run_checks(context)
+        result.raise_for_status()
+
+        response = json.loads(result.text)
+
+        name = response["name"]
+        files = response["files"]
+        outline = response["outline"]
+        warnings = response["warnings"] if "warnings" in response else []
+
+        from ansible_ai_connect.ai.apps import AiConfig
+
+        ai_config = cast(AiConfig, apps.get_app_config("ai"))
+        if ansible_lint_caller := ai_config.get_ansible_lint_caller():
+            for file in files:
+                file["content"] = ansible_lint_caller.run_linter(file["content"])
+
+        return name, files, outline, warnings
 
 
 class WCABasePlaybookExplanationPipeline(
@@ -587,7 +655,7 @@ class WCABasePlaybookExplanationPipeline(
                 f"{self.config.inference_url}/v1/wca/explain/ansible/playbook",
                 headers=headers,
                 json=data,
-                verify=self.config.verify_ssl,
+                verify=False,
             )
 
         result = post_request()
@@ -646,7 +714,7 @@ class WCABaseRoleExplanationPipeline(
                 f"{self.config.inference_url}/v1/wca/codegen/ansible/roles/explain",
                 headers=headers,
                 json=data,
-                verify=self.config.verify_ssl,
+                verify=False,
             )
 
         result = post_request()
